scripts/requirements.txt
# Google Drive automation dependencies
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
openpyxl>=3.1.0
